Switch to sapiensId for user entity identification across services
Replaced various IDs with sapiensId to standardize user entity identification in multiple services and DTOs. Adjusted related repositories, services, and validation to align with the new identifier structure.

+++ src/main/java/br/gov/agu/nutec/msescala/repository/AvaliadorRepository.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface AvaliadorRepository extends JpaRepository<AvaliadorEntity, Integer> {
+public interface AvaliadorRepository extends JpaRepository<AvaliadorEntity, Long> {
 
     Page<AvaliadorEntity> findByNomeContainingIgnoreCase(String nome, Pageable pageable);
